third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-array-argument-expected.txt

Validates that lockOrientation() can take an array as argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation is "portrait-primary"
locking orientation to ['portrait-primary', 'landscape-primary'].
PASS screen.lockOrientation(['portrait-primary', 'landscape-primary']) is true
PASS successfullyParsed is true

TEST COMPLETE
